pyrrhos/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
filelock==3.0.12
flake8==3.8.4
gdown==3.12.2
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.20.1
Pillow==8.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
pypandoc==1.5
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
tqdm==4.58.0
urllib3==1.26.3
